+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to the Node [default version scheme](https://docs.npmjs.com
 
 ### Changed
 
+- Public: Removed `userConsent` step option. User Consent is controlled by Onfido API.
 - Internal: Set `acceptSslCerts` to `true` for browserstack CI test to avoid "This connection is not private" screens
 - Internal: Migrate `ModalApp` to typescript
 - Internal: Add support for `.env` file